requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.0.0
Flask==1.0.2
idna==2.7
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.0
openapi-spec-validator==0.2.4
pymongo==3.7.2
python-dateutil==2.7.5
PyYAML==3.13
requests==2.20.0
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
wincertstore==0.2
